+++ spec/project/project_helper_spec.rb
@@ -10,7 +10,7 @@ module ProjectSpecs
 
     describe 'Targets' do
       it 'creates a new target' do
-        target = @helper.new_target(@project, :static_library, 'Pods', :ios, '6.0', @project.products_group)
+        target = @helper.new_target(@project, :static_library, 'Pods', :ios, '6.0', @project.products_group, :objc)
         target.name.should == 'Pods'
         target.product_type.should == 'com.apple.product-type.library.static'
 
@@ -28,12 +28,20 @@ module ProjectSpecs
       end
 
       it 'uses default build settings for Release and Debug configurations' do
-        target = @helper.new_target(@project, :static_library, 'Pods', :ios, '6.0', @project.products_group)
+        target = @helper.new_target(@project, :static_library, 'Pods', :ios, '6.0', @project.products_group, :objc)
         debug_settings = @helper.common_build_settings(:debug, :ios, '6.0', :static_library)
         release_settings = @helper.common_build_settings(:release, :ios, '6.0', :static_library)
 
         target.build_settings('Debug').should == debug_settings
         target.build_settings('Release').should == release_settings
+      end
+
+      it 'uses build settings for Swift language if required' do
+        target = @helper.new_target(@project, :framework, 'Pods', :ios, '8.0', @project.products_group, :objc)
+        target.build_settings('Debug')['SWIFT_OPTIMIZATION_LEVEL'].should.be.nil
+
+        target = @helper.new_target(@project, :framework, 'Pods', :ios, '8.0', @project.products_group, :swift)
+        target.build_settings('Debug')['SWIFT_OPTIMIZATION_LEVEL'].should == '-Onone'
       end
 
       it 'creates a new resources bundle' do